Added commented-out pip install cell and Markdwon instructions.
When running on Colab, it is necessary to pip install the required
dependencies. Therefore added a new, cleaner cell for pip-intalling the
required dependencies, and a cell with instructions for when to use this,
and what to do instead when running locally (`pip install -r requirements.txt`,
or `conda install --file requirements.txt`).